Exctracted job struct to own file
